requirements.txt
aiogram==3.17.0
asyncpg==0.30.0
aiosqlite==0.20.0
crypto-pay-api-sdk
flask==3.0.3
python-dotenv==1.0.0
aiohttp==3.11.18